CXP-1100: move GetWizardDataAction in internal api
